Boss Remake
-Remade how boss fights work

+++ Assets/Scripts/Boss/BossProjectile.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class BossProjectile : MonoBehaviour
 {
-    int speed = 1;
-    int timeAlive = 5;
+    float speed = 1;
+    float timeAlive = 5;
     float t = 0;
 
     void Update()
@@ -17,6 +17,6 @@ public class BossProjectile : MonoBehaviour
         transform.Translate(-speed * Time.deltaTime, 0, 0);
     }
 
-    public void SetSpeed(int s) { speed = s; }
-    public void SetDistance(int d) { timeAlive= d; }
+    public void SetSpeed(float s) { speed = s; }
+    public void SetDistance(float d) { timeAlive= d; }
 }
